Revert "Convert data type of `DistanceSpacing` to float"
This reverts commit 7aaa88cac2be239dcc778a1082585f8e8d5a7e55.

+++ osu.Game.Tests/Editing/TestSceneHitObjectComposerDistanceSnapping.cs
@@ -222,7 +222,7 @@ namespace osu.Game.Tests.Editing
         {
             public new EditorBeatmap EditorBeatmap => base.EditorBeatmap;
 
-            public new Bindable<float> DistanceSpacingMultiplier => base.DistanceSpacingMultiplier;
+            public new Bindable<double> DistanceSpacingMultiplier => base.DistanceSpacingMultiplier;
 
             public TestHitObjectComposer()
                 : base(new OsuRuleset())
